requirements.txt
Django==5.2.7
google-generativeai==0.8.5
python-dotenv==1.1.1
PyMuPDF==1.26.5
psycopg2-binary==2.9.11
openai==2.5.0
celery==5.5.3
redis==6.4.0